+++ test/unit/c11_test_shmem_atomic_or.c
@@ -30,6 +30,7 @@
  * SOFTWARE.
  */
 
+#include <stddef.h>
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdbool.h>
@@ -51,7 +52,7 @@ enum op { OR = 0, CTX_OR, FETCH_OR, CTX_FETCH_OR };
     const int mype = shmem_my_pe();                                     \
     const int npes = shmem_n_pes();                                     \
     TYPE old = (TYPE)0;                                                 \
-    if (npes-1 > sizeof(TYPE)) break; /* Avoid overflow */              \
+    if ((size_t) npes-1 > sizeof(TYPE)) break; /* Avoid overflow */     \
     for (int i = 0; i < npes; i++)                                      \
       switch (OP) {                                                     \
         case OR:                                                        \
