+++ test/unit/cxx_test_shmem_atomic_and.cpp
@@ -30,6 +30,7 @@
  * SOFTWARE.
  */
 
+#include <stddef.h>
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdbool.h>
@@ -47,7 +48,7 @@ enum op { AND = 0, CTX_AND, FETCH_AND, CTX_FETCH_AND };
   do {                                                                  \
     static TYPE remote = ~(TYPE)0;                                      \
     TYPE old = (TYPE)0;                                                 \
-    if (npes-1 > sizeof(TYPE)) break; /* Avoid overflow */              \
+    if ((size_t) npes-1 > sizeof(TYPE)) break; /* Avoid overflow */     \
     for (int i = 0; i < npes; i++)                                      \
       switch (OP) {                                                     \
         case AND:                                                       \
